backend/requirements.txt
# FastAPI and ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Object Storage (MinIO)
minio==7.2.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Utilities
python-dateutil==2.8.2
python-md5==1.0.0
aiofiles==23.2.1

# Excel Export
openpyxl==3.1.2

# Logging
loguru==0.7.2

# Task Queue
celery==5.3.6
